+++ services/src/test/java/org/apache/druid/server/AsyncQueryForwardingServletTest.java
@@ -652,7 +652,7 @@ public class AsyncQueryForwardingServletTest extends BaseJettyTest
         null,
         null,
         stubServiceEmitter,
-        new NoopRequestLogger(),
+        NoopRequestLogger.instance(),
         new DefaultGenericQueryMetricsFactory(),
         new AuthenticatorMapper(ImmutableMap.of()),
         properties,
@@ -762,8 +762,8 @@ public class AsyncQueryForwardingServletTest extends BaseJettyTest
               hostFinder,
               injector.getProvider(HttpClient.class),
               injector.getInstance(DruidHttpClientConfig.class),
-              new NoopServiceEmitter(),
-              new NoopRequestLogger(),
+              NoopServiceEmitter.instance(),
+              NoopRequestLogger.instance(),
               new DefaultGenericQueryMetricsFactory(),
               new AuthenticatorMapper(ImmutableMap.of()),
               new Properties(),
@@ -984,8 +984,8 @@ public class AsyncQueryForwardingServletTest extends BaseJettyTest
             null,
             null,
             null,
-            new NoopServiceEmitter(),
-            new NoopRequestLogger(),
+            NoopServiceEmitter.instance(),
+            NoopRequestLogger.instance(),
             new DefaultGenericQueryMetricsFactory(),
             new AuthenticatorMapper(ImmutableMap.of()),
             new Properties(),
